BackEnd/requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
bitsandbytes
Bottleneck
Brotli
certifi
cffi
click
crefi
cryptography
datasets
dill
dnspython
faiss-cpu
fastapi
filelock
frozenlist
fsspec
h11
huggingface_hub
idna
Jinja2
joblib
logger
MarkupSafe
mpmath
multidict
multiprocess
networkx
newsapi-python
numexpr
numpy
packaging
pandas
pillow
propcache
protobuf
psutil
pyarrow
pycparser
pydantic
pydantic-core
pydantic-settings
pymongo
PySocks
python-dateutil
python-dotenv
pytz
pyxattr
PyYAML
regex
requests
safetensors
scikit-learn
scipy
sentence-transformers
sentencepiece
setuptools
six
sniffio
starlette
sympy
threadpoolctl
tokenizers
torch
tqdm
transformers
typing_extensions
tzdata
urllib3
uvicorn
wheel
xxhash
yarl
